refactor: move SHE helpers into src/core/lib/SHE.mjs
The key slot table, protection flags, memory update constants, SHE KDF
and the AES primitives it uses now live in src/core/lib/SHE.mjs, leaving
the operation to argument handling and message assembly. The local hex
parser is replaced by fromHex from lib/Hex.mjs.
